requirements.txt
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
defusedxml==0.7.1
dj-rest-auth==2.2.3
Django==4.0.3
django-allauth==0.49.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
idna==3.3
lxml==4.8.0
mysqlclient==2.1.0
oauthlib==3.2.0
packaging==21.3
pikepdf==5.0.1
Pillow==9.0.1
pycparser==2.21
pycryptodome==3.14.1
PyJWT==2.3.0
pyparsing==3.0.7
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
urllib3==1.26.8
